(U2C #18) move EventPublisher & FeatureRequester out of main namespace (#119)
